+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,3 @@
 *   Keyed NotInParallelAttribute
+- Throw DependencyFailedException if a dependency fails
+- Better timing out of modules if the pipeline has failed